requirements.txt
flask==2.0.1
line-bot-sdk==2.3.0
openai==1.1.1
python-dotenv==0.19.2
redis==4.4.0
flask-limiter==2.8.1
mysql-connector-python==8.0.27
waitress==2.1.2
gunicorn==20.1.0
regex==2022.10.31
APScheduler==3.9.1
requests==2.28.1